requirements.txt
requests>=2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.3
schedule>=1.2.0
arxiv>=2.1.0
